fix: standardize factoring and closest vector inputs
Derive factoring widths from the target and canonicalize factor witnesses. Model CVP without artificial coefficient bounds and register exact sphere enumeration. Make QUBO numeric variants and same-problem variant reductions explicit so the standard models remain executable.

+++ src/unit_tests/example_db.rs
@@ -1211,7 +1211,7 @@ fn test_find_rule_example_paintshop_to_qubo() {
     };
     let target = ProblemRef {
         name: "QUBO".to_string(),
-        variant: BTreeMap::from([("weight".to_string(), "f64".to_string())]),
+        variant: BTreeMap::from([("weight".to_string(), "i64".to_string())]),
     };
     let example = find_rule_example(&source, &target).unwrap();
     assert_eq!(example.source.problem, "PaintShop");
